requirements.txt
google-generativeai
pyserial
opencv-python
Pillow